+++ lib/rubocop/cop/layout/empty_lines_around_access_modifier.rb
@@ -5,7 +5,7 @@ module RuboCop
     module Layout
       # Access modifiers should be surrounded by blank lines.
       #
-      # @example
+      # @example EnforcedStyle: around (default)
       #
       #   # bad
       #   class Foo
@@ -22,12 +22,34 @@ module RuboCop
       #
       #     def baz; end
       #   end
+      #
+      # @example EnforcedStyle: only_before
+      #
+      #   # bad
+      #   class Foo
+      #     def bar; end
+      #     private
+      #     def baz; end
+      #   end
+      #
+      #   # good
+      #   class Foo
+      #     def bar; end
+      #
+      #     private
+      #     def baz; end
+      #   end
+      #
       class EmptyLinesAroundAccessModifier < Cop
+        include ConfigurableEnforcedStyle
         include RangeHelp
 
         MSG_AFTER = 'Keep a blank line after `%<modifier>s`.'
         MSG_BEFORE_AND_AFTER = 'Keep a blank line before and after ' \
                                '`%<modifier>s`.'
+
+        MSG_BEFORE_FOR_ONLY_BEFORE = 'Keep a blank line before `%<modifier>s`.'
+        MSG_AFTER_FOR_ONLY_BEFORE = 'Remove a blank line after `%<modifier>s`.'
 
         def initialize(config = nil, options = nil)
           super
@@ -63,7 +85,12 @@ module RuboCop
         def on_send(node)
           return unless node.bare_access_modifier?
 
-          return if empty_lines_around?(node)
+          case style
+          when :around
+            return if empty_lines_around?(node)
+          when :only_before
+            return if allowed_only_before_style?(node)
+          end
 
           add_offense(node)
         end
@@ -76,13 +103,34 @@ module RuboCop
               corrector.insert_before(line, "\n")
             end
 
-            unless next_line_empty?(node.last_line)
-              corrector.insert_after(line, "\n")
-            end
+            correct_next_line_if_denied_style(corrector, node, line)
           end
         end
 
         private
+
+        def allowed_only_before_style?(node)
+          if node.special_modifier?
+            return false if next_line_empty?(node.last_line)
+          end
+
+          previous_line_empty?(node.first_line)
+        end
+
+        def correct_next_line_if_denied_style(corrector, node, line)
+          case style
+          when :around
+            unless next_line_empty?(node.last_line)
+              corrector.insert_after(line, "\n")
+            end
+          when :only_before
+            if next_line_empty?(node.last_line)
+              range = next_empty_line_range(node)
+
+              corrector.remove(range)
+            end
+          end
+        end
 
         def previous_line_ignoring_comments(processed_source, send_line)
           processed_source[0..send_line - 2].reverse.find do |line|
@@ -128,7 +176,20 @@ module RuboCop
           line == @class_or_module_def_last_line - 1
         end
 
+        def next_empty_line_range(node)
+          source_range(processed_source.buffer, node.last_line + 1, 0)
+        end
+
         def message(node)
+          case style
+          when :around
+            message_for_around_style(node)
+          when :only_before
+            message_for_only_before_style(node)
+          end
+        end
+
+        def message_for_around_style(node)
           send_line = node.first_line
 
           if block_start?(send_line) ||
@@ -136,6 +197,16 @@ module RuboCop
             format(MSG_AFTER, modifier: node.loc.selector.source)
           else
             format(MSG_BEFORE_AND_AFTER, modifier: node.loc.selector.source)
+          end
+        end
+
+        def message_for_only_before_style(node)
+          modifier = node.loc.selector.source
+
+          if next_line_empty?(node.last_line)
+            format(MSG_AFTER_FOR_ONLY_BEFORE, modifier: modifier)
+          else
+            format(MSG_BEFORE_FOR_ONLY_BEFORE, modifier: modifier)
           end
         end
       end